+++ metac_alloc_node.h
@@ -11,10 +11,10 @@ metac_expression_t* AllocNewExpression(metac_expression_kind_t kind);
 #define AllocNewDeclaration(KIND, RESULT_PTR) \
     (KIND ## _t*) AllocNewDeclaration_(KIND, sizeof(KIND ##_t), ((void**)(RESULT_PTR)), __LINE__)
 
-metac_declaration_t* AllocNewDeclaration_(metac_declaration_kind_t kind, size_t nodeSize, void** result_ptr, uint32_t line);
+metac_declaration_t* AllocNewDeclaration_(metac_declaration_kind_t kind, uint32_t nodeSize, void** result_ptr, uint32_t line);
 
 
 #define AllocNewStatement(KIND, RESULT_PTR) \
     (KIND ## _t*) AllocNewStatement_(KIND, sizeof(KIND ##_t), ((void**)(RESULT_PTR)))
 
-metac_statement_t* AllocNewStatement_(metac_statement_kind_t kind, size_t nodeSize, void** result_ptr);
+metac_statement_t* AllocNewStatement_(metac_statement_kind_t kind, uint32_t nodeSize, void** result_ptr);
